fix(registry): 🐛 enforce minimum ring depth and detection-time timestamps
## Summary

Addresses two review findings on PR #11. Ring depth of 0 is now rejected
at both CLI validation and runtime assertion. Timestamps are now assigned
at turn detection time (in the PTY wrapper) instead of broker receipt time,
per CONTRACT_REGISTRY.md §73.

## Highlights

- CLI `--ring-depth` uses `value_parser!(u64).range(1..)` to reject 0
- `TurnRingBuffer::new()` panics on capacity 0 (defensive invariant)
- `Turn` struct carries `timestamp: u64` set via `epoch_millis()` at detection
- `Message::TurnCompleted` gains `timestamp: u64` field on the wire
- `TurnRingBuffer::push()` accepts caller-supplied timestamp, no longer calls `SystemTime::now()`
- Full chain: detector → Turn → TurnCompleted → handler → store_turn → ring.push

+++ src/main.rs
@@ -31,7 +31,7 @@ async fn main() {
             max_turn_size,
         } => {
             let config = broker::state::RingConfig {
-                depth: ring_depth,
+                depth: ring_depth as usize,
                 max_turn_bytes: max_turn_size,
             };
             if let Err(e) = broker::run(config).await {
